+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/init-stack
 
+## 2.7.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes