+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/stack-ui
 
+## 2.7.14
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.14
+
 ## 2.7.13
 
 ### Patch Changes